index canbios lab y examen

+++ models/postgres/labs.js
@@ -88,9 +88,9 @@ export class LabModel {
   static async updateImg ({ id, input }) {
     try {
       // eslint-disable-next-line camelcase
-      const abatar = input
+      const logo = input
 
-      console.log('entro en model udate img: ' + abatar + id)
+      /*      console.log('entro en model udate img: ' + abatar + id)
 
       const resultUrlDelete = await conn.query('SELECT users_abatar FROM users WHERE users_id = $1;', [id])
 
@@ -99,14 +99,19 @@ export class LabModel {
       const abatarUrlDelete = resultUrlDelete.rows[0].users_abatar
       const nombreImgDelete = nombreFinalImagenByUrl(abatarUrlDelete)
       console.log(abatarUrlDelete)
-      console.log(nombreImgDelete)
+      console.log(nombreImgDelete) */
 
+      const resultUrlDelete = await conn.query('SELECT lab_logo FROM lab WHERE lab_id = $1;', [id])
+
+      if (resultUrlDelete.row === 0) return resultUrlDelete.row
+
+      const nombreImgDelete = nombreFinalImagenByUrl(resultUrlDelete.rows[0].lab_logo)
       if (nombreImgDelete !== 'default.jpg') {
         console.log('BOORANDO ARCHIVO')
         await borrarImagen(nombreImgDelete)
       }
 
-      const result = await conn.query('UPDATE users SET  users_abatar=$1 WHERE users_id=$2 RETURNING *;', [abatar, id])
+      const result = await conn.query('UPDATE lab SET lab_logo=$1 WHERE lab_id=$2 RETURNING *;', [logo, id])
       return result.rows
     } catch (e) {
       console.log('Error DB en Users IMG update')
